requirements.txt
aiohttp>=3.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0
structlog>=23.2.0
click>=8.1.0
psutil>=5.9.0
google-auth>=2.23.0
